Undid move of SocketAPI and TransactionManager to their own folder to make diffing easier on the PR.
